同步作者代码

+++ server/FFmpegSource.cpp
@@ -59,6 +59,11 @@ static bool is_local_ip(const string &ip){
     return false;
 }
 
+void FFmpegSource::setupRecord(bool enable_hls, bool enable_mp4){
+    _enable_hls = enable_hls;
+    _enable_mp4 = enable_mp4;
+}
+
 void FFmpegSource::play(const string &src_url,const string &dst_url,int timeout_ms,const onPlay &cb) {
     GET_CONFIG(string,ffmpeg_bin,FFmpeg::kBin);
     GET_CONFIG(string,ffmpeg_cmd,FFmpeg::kCmd);
@@ -197,8 +202,8 @@ void FFmpegSource::startTimer(int timeout_ms) {
             strongSelf->findAsync(0, [&](const MediaSource::Ptr &src) {
                 //同步查找流
                 if (!src) {
-                    //流不在线，重新拉流
-                    if(strongSelf->_replay_ticker.elapsedTime() > 10 * 1000){
+                    //流不在线，重新拉流, 这里原先是10秒超时，实际发现10秒不够，改成20秒了
+                    if(strongSelf->_replay_ticker.elapsedTime() > 20 * 1000){
                         //上次重试时间超过10秒，那么再重试FFmpeg拉流
                         strongSelf->_replay_ticker.resetTime();
                         strongSelf->play(strongSelf->_src_url, strongSelf->_dst_url, timeout_ms, [](const SockException &) {});
@@ -233,7 +238,7 @@ void FFmpegSource::setOnClose(const function<void()> &cb){
 }
 
 bool FFmpegSource::close(MediaSource &sender, bool force) {
-    auto listener = _listener.lock();
+    auto listener = getDelegate();
     if(listener && !listener->close(sender,force)){
         //关闭失败
         return false;
@@ -245,18 +250,30 @@ bool FFmpegSource::close(MediaSource &sender, bool force) {
     return true;
 }
 
+MediaOriginType FFmpegSource::getOriginType(MediaSource &sender) const{
+    return MediaOriginType::ffmpeg_pull;
+}
+
+string FFmpegSource::getOriginUrl(MediaSource &sender) const{
+    return _src_url;
+}
+
+std::shared_ptr<SockInfo> FFmpegSource::getOriginSock(MediaSource &sender) const {
+    return nullptr;
+}
+
 void FFmpegSource::onGetMediaSource(const MediaSource::Ptr &src) {
-    auto listener = src->getListener();
+    auto listener = src->getListener(true);
     if (listener.lock().get() != this) {
-        //防止多次进入onGetMediaSource函数导致无效递归调用的bug
-        _listener = listener;
+        //防止多次进入onGetMediaSource函数导致无限递归调用的bug
+        setDelegate(listener);
         src->setListener(shared_from_this());
-    } else {
-        WarnL << "多次触发onGetMediaSource事件:"
-              << src->getSchema() << "/"
-              << src->getVhost() << "/"
-              << src->getApp() << "/"
-              << src->getId();
+        if (_enable_hls) {
+            src->setupRecord(Recorder::type_hls, true, "");
+        }
+        if (_enable_mp4) {
+            src->setupRecord(Recorder::type_mp4, true, "");
+        }
     }
 }
 
@@ -264,20 +281,27 @@ void FFmpegSnap::makeSnap(const string &play_url, const string &save_path, float
     GET_CONFIG(string,ffmpeg_bin,FFmpeg::kBin);
     GET_CONFIG(string,ffmpeg_snap,FFmpeg::kSnap);
     GET_CONFIG(string,ffmpeg_log,FFmpeg::kLog);
-
-    std::shared_ptr<Process> process = std::make_shared<Process>();
-    auto delayTask = EventPollerPool::Instance().getPoller()->doDelayTask(timeout_sec * 1000,[process,cb](){
-        if(process->wait(false)){
-            //FFmpeg进程还在运行，超时就关闭它
-            process->kill(2000);
+    Ticker ticker;
+    WorkThreadPool::Instance().getPoller()->async([timeout_sec, play_url,save_path,cb, ticker](){
+        auto elapsed_ms = ticker.elapsedTime();
+        if (elapsed_ms > timeout_sec * 1000) {
+            //超时，后台线程负载太高，当代太久才启动该任务
+            cb(false);
+            return;
         }
-        return 0;
-    });
-
-    WorkThreadPool::Instance().getPoller()->async([process,play_url,save_path,delayTask,cb](){
         char cmd[1024] = {0};
         snprintf(cmd, sizeof(cmd),ffmpeg_snap.data(),ffmpeg_bin.data(),play_url.data(),save_path.data());
+        std::shared_ptr<Process> process = std::make_shared<Process>();
         process->run(cmd,ffmpeg_log.empty() ? "" : File::absolutePath("",ffmpeg_log));
+        //定时器延时应该减去后台任务启动的延时
+        auto delayTask = EventPollerPool::Instance().getPoller()->doDelayTask(timeout_sec * 1000 - elapsed_ms,[process,cb](){
+            if(process->wait(false)){
+                //FFmpeg进程还在运行，超时就关闭它
+                process->kill(2000);
+            }
+            return 0;
+        });
+
         //等待FFmpeg进程退出
         process->wait(true);
         //FFmpeg进程退出了可以取消定时器了
